--- src/components/appMenuSections/dashboard/styles.css
@@ -1,3 +1,0 @@
-.page-heading-title {
-  /* color: #2f2f2f; */
-}
