--- src/components/appMenuSections/profile/styles.css
@@ -1,8 +1,0 @@
-.default-page-container {
-  height: 100vh;
-  /* width: 100vw; - Do not add this, because if there's vertical scroll, the page will also have horizontal scroll. without it,
-                     the page width will be automatically adjusted for the contents. You only need the height. */
-  /* background-color: #eeeeee; */
-  /* color: #2f2f2f; */
-  /* color: #121212; */
-}
