--- src/components/appMenuSections/style.css
@@ -1,9 +1,0 @@
-.minerProfileButtons {
-  margin-left: 10px !important;
-  /* margin-bottom: 20px !important; */
-  /* display: block !important; */
-}
-
-.liMenuMiningPool {
-  display: block !important;
-}
