+++ src/css/navbars/styles.css
@@ -4,14 +4,9 @@
 }
 
 /* loginButtonIconHeader */
-button.Connect .css-i4bv87-MuiSvgIcon-root {
+/* button.Connect .css-i4bv87-MuiSvgIcon-root {
   color: #f9b11c !important;
-}
-
-#loginIconHeader {
-  color: green !important;
-  background-color: blue !important;
-}
+} */
 
 /* switch theme icon button light theme */
 .css-decirt-MuiSwitch-root .MuiSwitch-thumb {
@@ -29,6 +24,7 @@ button.Connect .css-i4bv87-MuiSvgIcon-root {
   color: #f9b11c !important;
 }
 
+/* show the current section in the side bard menu */
 .active-custom .css-i4bv87-MuiSvgIcon-root,
 .active-custom .css-tlelie-MuiListItemText-root {
   color: #f9b11c !important;
@@ -39,14 +35,15 @@ button.Connect .css-i4bv87-MuiSvgIcon-root {
   color: #f9b11c !important;
 }
 
-/* .MuiListItem-root .MuiListItem-gutters .css-1nxmd3h-MuiListItem-root .css-i4bv87-MuiSvgIcon-root {
-  background-color: red !important;
-  color: blue !important;
-} */
-
-.css-5ryogn-MuiButtonBase-root-MuiSwitch-switchBase.Mui-checked .css-8je8zh-MuiTouchRipple-root {
-  /* background-color: #f9b11c !important;
-  color: white !important; */
+.default-page-container {
+  height: 100vh;
+  /* width: 100vw; - Do not add this, because if there's vertical scroll, the page will also have horizontal scroll. without it,
+                     the page width will be automatically adjusted for the contents. You only need the height. */
+  /* background-color: #eeeeee; */
+  /* color: #2f2f2f; */
+  /* color: #121212; */
 }
 
-/* .css-i4bv87-MuiSvgIcon-root */
+.liMenuMiningPool {
+  display: block !important;
+}
